requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-secret-manager==2.18.1
google-cloud-storage==2.10.0
google-api-python-client==2.108.0
